WIP: split code
