+++ example/example-update.js
@@ -1,0 +1,32 @@
+'use strict';
+
+const FauxMo = require('../lib');
+
+const devices = {
+  devices: [
+    {
+      name: 'office light',
+      port: 11000,
+      handler: (action) => {
+        console.log('office light action:', action);
+      }
+    },
+    {
+      name: 'office fan',
+      port: 11001,
+      handler: (action) => {
+        console.log('office fan action:', action);
+      }
+    }
+  ]
+};
+
+let fauxMo = new FauxMo();
+console.log('started..');
+
+fauxMo.updateDevices(devices);
+console.log('updated..');
+
+fauxMo.stopServer();
+fauxMo.stopServer(100);
+console.log('stop..');
